pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = ""
authors = [
    {name = "Vitor Brito",email = "joaovitor16611661@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "django (>=5.1.7,<6.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "django-debug-toolbar (>=5.0.1,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "flake8 (>=7.1.2,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
